Compress images
It's taking way too long for people to download our LFS data. This
appears to be caused by a lot of uncompressed images. This commit
compresses the images and zips containing the images to make LFS faster.

Finally, this commit modifies the image compressor so that it can work
on .JPG and .JPEG files (as well as their lowercase counterparts). It
turns out some files from our cameras used these fully-uppercase
spellings of the extension.